LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/filter/offscreencanvas.filter.w-expected.txt

CONSOLE MESSAGE: line 4: ReferenceError: Can't find variable: OffscreenCanvas

Harness Error (FAIL), message = ReferenceError: Can't find variable: OffscreenCanvas

TIMEOUT offscreencanvas Test timed out

Harness Error (FAIL), message = ReferenceError: Can't find variable: OffscreenCanvas

TIMEOUT offscreencanvas Test timed out
